// app/utils/analytics/types.ts
export interface UserProperties {
  DOB: string
  active_plan: string[]
  age: number
  allow_att_tracking: boolean
  aum: number
  bvn_verified: boolean
  cio_subscription_preferences: string
  country: string
  email_verified: boolean
  employment_status: string
  gender: string
  has_autoinvest_on: boolean
  has_bank_account: boolean
  has_reinvest_on: boolean
  id_proof_uploaded: boolean
  id_upload_type: string[]
  id_verified: boolean
  in_app_distribution: string
  inapp_attribution_channel: string
  is_affiliate: boolean
  is_referred: string | boolean
  job_sector: string
  login_methods: string
  marital_status: string
  monthly_income_range: string
  phone_number: string
  plan_types: string[]
  push_notification_preference: 'allow' | 'disallow'
  referred_investors: number
  referred_signups: number
  signup_date: string
  user_id: string
  verification: string[]
  verification_type_done: string
  virtual_acct_generated: boolean
}
// ...
export type EventName = keyof EventLogData

export type RequireAtLeastOne<T, Keys extends keyof T = keyof T> = Pick<
  T,
  Exclude<keyof T, Keys>
> &
  {
    [K in Keys]-?: Required<Pick<T, K>> & Partial<Pick<T, Exclude<Keys, K>>>
  }[Keys]